# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LogB",
    "version": "0.1.0",
    "description": "The HTTP API behind LogB, served under `/api`.\n\nTwo credentials work everywhere except issuing and listing tokens: the `logb_session` cookie a browser gets from `POST /auth/login`, and an `Authorization: Bearer logb_pat_...` token for clients that are not browsers. A client obtains its first token by logging in interactively once and calling `POST /auth/tokens`; issuing and listing tokens deliberately refuse bearer credentials, so a leaked token cannot mint replacements. `DELETE /auth/tokens/{id}` is the one exception: a bearer token may revoke itself, so an app can sign out without ever holding a cookie -- but not any other token. `GET /health` lists `token-self-revoke` in its `features` array so a client can detect support before relying on it.\n\n**Writing offline.** `POST /objects/{id}/activities` and `POST /objects/{id}/attachments` accept a `client_op_id`: any string unique to the write, generated once by the client and kept across every retry. The server records it under a unique index, so replaying a write whose response was lost resolves to the row already created instead of duplicating it. Any client that queues writes while offline should use it.\n\nThis document is checked against the running router by `tests/openapi.rs`: a route added, removed or renamed without updating it fails the build."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "security": [
    {
      "sessionCookie": []
    },
    {
      "bearerToken": []
    }
  ],
  "paths": {
...
    "/me/notifications": {
      "get": {
        "summary": "Where the caller's daily digest goes, and what a browser needs to subscribe to push.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationSettings"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Set the caller's own digest webhook and its format.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotificationSettingsInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationSettings"
                }
              }
            }
          }
        }
      }
    },
...
    "/push/subscriptions": {
      "post": {
        "summary": "Subscribe a browser to push notifications. An endpoint already known moves to the caller.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushSubscription"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Subscribed"
          }
        }
      },
      "delete": {
        "summary": "Unsubscribe one of the caller's browsers. Idempotent.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "endpoint"
                ],
                "properties": {
                  "endpoint": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Unsubscribed"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "logb_session",
        "description": "Set by `POST /auth/login`. HttpOnly, SameSite=Lax, 30 days."
      },
      "bearerToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "An API token from `POST /auth/tokens`. Does not expire; revoke it to end it."
      }
    },
    "schemas": {
...
      "NotificationSettings": {
        "type": "object",
        "properties": {
          "url": {
            "type": [
              "string",
              "null"
            ],
            "description": "This user's own digest webhook. Null keeps them in the instance digest"
          },
          "format": {
            "type": "string",
            "enum": [
              "text",
              "json"
            ]
          },
          "push_devices": {
            "type": "integer",
            "description": "Browsers this user has turned push notifications on in"
          },
          "vapid_public_key": {
            "type": "string",
            "description": "`applicationServerKey` for PushManager.subscribe: an uncompressed P-256 point, base64url"
          },
          "instance_webhook": {
            "type": "boolean",
            "description": "Whether LOGB_NOTIFY_URL is set"
          },
          "hour": {
            "type": "integer",
            "description": "The hour, in the instance timezone, the digest goes out"
          }
        }
      },
      "NotificationSettingsInput": {
        "type": "object",
        "properties": {
          "url": {
            "type": [
              "string",
              "null"
            ],
            "description": "An http(s) URL; blank or null for none"
          },
          "format": {
            "type": "string",
            "enum": [
              "text",
              "json"
            ],
            "default": "text"
          }
        }
      },
...
      "PushSubscription": {
        "type": "object",
        "required": [
          "endpoint",
          "keys"
        ],
        "description": "What PushSubscription.toJSON() produces. The endpoint must be https",
        "properties": {
          "endpoint": {
            "type": "string"
          },
          "keys": {
            "type": "object",
            "required": [
              "p256dh",
              "auth"
            ],
            "properties": {
              "p256dh": {
                "type": "string"
              },
              "auth": {
                "type": "string"
              }
            }
          }
        }
      },
...
    }
  }
}